+++ sources/shapes/Triangle.cpp
@@ -2,7 +2,7 @@
 
 Triangle::Triangle()
 {
-    float vertices[] = {
+    double vertices[] = {
         -0.5f, -0.5f, 0.0f, 0.0f, 0.0f, 1.0f, 
          0.5f, -0.5f, 0.0f, 0.0f, 0.0f, 1.0f, 
          0.0f,  0.5f, 0.0f, 0.0f, 0.0f, 1.0f 
@@ -16,10 +16,10 @@ Triangle::Triangle()
 
     glBufferData(GL_ARRAY_BUFFER, sizeof(vertices), vertices, GL_STATIC_DRAW);
 
-    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 6 * sizeof(float), (void*)0);
+    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 6 * sizeof(double), (void*)0);
     glEnableVertexAttribArray(0);
 
-    glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, 6 * sizeof(float), (void*)(3 * sizeof(float)));
+    glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, 6 * sizeof(double), (void*)(3 * sizeof(double)));
     glEnableVertexAttribArray(1);
 
     glBindVertexArray(0);
